fix: Broken tests
